fix bugs

+++ src/main/java/com/example/SwipeRest/controller/ApartmentController.java
@@ -2,6 +2,8 @@ package com.example.SwipeRest.controller;
 
 import com.example.SwipeRest.dto.ApartmentDTO;
 import com.example.SwipeRest.service.impl.ApartmentServiceImpl;
+import com.example.SwipeRest.service.impl.LCDServiceImpl;
+import com.example.SwipeRest.service.impl.UserServiceImpl;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -20,6 +22,8 @@ import org.springframework.web.bind.annotation.*;
 @Tag(name = "Apartment")
 public class ApartmentController {
     private final ApartmentServiceImpl apartmentService;
+    private final LCDServiceImpl lcdService;
+    private final UserServiceImpl userService;
     private Logger log = LoggerFactory.getLogger(ApartmentController.class);
     @Operation(summary = "Get all apartment")
     @ApiResponses({
@@ -41,6 +45,9 @@ public class ApartmentController {
     })
     @GetMapping("/{id}")
     public ResponseEntity findByIdApartment(@PathVariable @Schema(example = "2")int id){
+        if (id<0){
+            return ResponseEntity.badRequest().body("Id cannot be negative");
+        }
         ApartmentDTO apartmentDTO = apartmentService.findByIdDTO(id);
         if (apartmentDTO!=null) {
             log.info("Request find apartment "+id);
@@ -60,6 +67,14 @@ public class ApartmentController {
     @PostMapping("/add")
     public ResponseEntity saveApartment(@RequestBody @Valid ApartmentDTO apartmentDTO){
         log.info("Request save apartment");
+        apartmentDTO.setId(0);
+        apartmentDTO.getPhotos().forEach(photoDTO -> {photoDTO.setId(0);});
+        if (lcdService.findById(apartmentDTO.getIdLcd()) == null){
+            return ResponseEntity.badRequest().body("idLcd wrong");
+        }
+        if (userService.findById(apartmentDTO.getClient()) == null){
+            return ResponseEntity.badRequest().body("idClient wrong");
+        }
         return ResponseEntity.ok("Success:\n"+apartmentService.saveDTO(apartmentDTO));
     }
     @Operation(summary = "Delete apartment by id")
@@ -70,11 +85,14 @@ public class ApartmentController {
     })
     @DeleteMapping("/delete/{id}")
     public  ResponseEntity deleteApartment(@PathVariable @Schema(example = "0") int id){
+        if (id<0){
+            return ResponseEntity.badRequest().body("Id cannot be negative");
+        }
         ApartmentDTO apartmentDTO = apartmentService.findByIdDTO(id);
         if (apartmentDTO!=null){
             log.info("Request delete apartment "+id);
             apartmentService.deleteById(id);
-            return ResponseEntity.ok("Success:\n"+apartmentDTO);
+            return ResponseEntity.ok("Success");
         }
 
         else {log.info("Apartment not found "+id);
@@ -89,6 +107,15 @@ public class ApartmentController {
     })
     @PutMapping("/update/{id}")
     public ResponseEntity updateApartment(@PathVariable @Schema(example = "2") int id, @RequestBody @Valid ApartmentDTO apartmentDTO){
+        if (id<0){
+            return ResponseEntity.badRequest().body("Id cannot be negative");
+        }
+        if (lcdService.findById(apartmentDTO.getIdLcd()) == null){
+            return ResponseEntity.badRequest().body("idLcd wrong");
+        }
+        if (userService.findById(apartmentDTO.getClient()) == null){
+            return ResponseEntity.badRequest().body("idClient wrong");
+        }
         ApartmentDTO apartmentDTO1 = apartmentService.findByIdDTO(id);
         if (apartmentDTO1!=null){
             log.info("Request update apartment "+id);
